extract MatchPledge to separate file
