+++ app/grants/models/__init__.py
@@ -59,10 +59,17 @@ from .grant import Grant, GrantCLR, GrantCLRCalculation, GrantType
 from .subscription import Subscription
 from .grant_collection import GrantCollection
 from .grant_api_key import GrantAPIKey
+from .match_pledge import MatchPledge
 
 logger = logging.getLogger(__name__)
 
 
+
+def next_month():
+    """Get the next month time."""
+    return localtime(timezone.now() + timedelta(days=30))
+
+    
 class Donation(SuperModel):
     """Define the structure of an optional donation. These donations are
        additional funds sent to Gitcoin as part of contributing or subscribing
@@ -145,55 +152,6 @@ class Donation(SuperModel):
         return f"id: {self.pk}; from:{profile.handle}; {tx_id} => ${token_amount_usdt}; {naturaltime(self.created_on)}"
 
 
-def next_month():
-    """Get the next month time."""
-    return localtime(timezone.now() + timedelta(days=30))
-
-
-class MatchPledge(SuperModel):
-    """Define the structure of a MatchingPledge."""
-
-    PLEDGE_TYPES = [
-        ('tech', 'tech'),
-        ('media', 'media'),
-        ('health', 'health'),
-        ('change', 'change')
-    ]
-
-    active = models.BooleanField(default=False, help_text=_('Whether or not the MatchingPledge is active.'))
-    profile = models.ForeignKey(
-        'dashboard.Profile',
-        related_name='matchPledges',
-        on_delete=models.CASCADE,
-        help_text=_('The MatchingPledgers profile.'),
-        null=True,
-    )
-    amount = models.DecimalField(
-        default=1,
-        decimal_places=4,
-        max_digits=50,
-        help_text=_('The matching pledge amount in DAI.'),
-    )
-    pledge_type = models.CharField(max_length=15, null=True, blank=True, choices=PLEDGE_TYPES, help_text=_('CLR pledge type'))
-    comments = models.TextField(default='', blank=True, help_text=_('The comments.'))
-    end_date = models.DateTimeField(null=False, default=next_month)
-    data = JSONField(null=True, blank=True)
-    clr_round_num = models.ForeignKey(
-        'grants.GrantCLR',
-        on_delete=models.CASCADE,
-        help_text=_('Pledge CLR Round.'),
-        null=True,
-        blank=True
-    )
-
-    @property
-    def data_json(self):
-        import json
-        return json.loads(self.data)
-
-    def __str__(self):
-        """Return the string representation of this object."""
-        return f"{self.profile} <> {self.amount} DAI"
 
 
 class PhantomFunding(SuperModel):
